tests/wpt/tests/mathml/relations/html5-tree/a-type-getter-001.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>MathMLAnchorElement.type getter</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <math>
      <a id="test" href="a">Example Link</a>
    </math>
  </body>
  <script>
      test(function() {
          var a = document.getElementById("test");
          assert_equals(a.type, "");
      }, "Test anchor's type getter when empty");
  </script>
</html>
