third_party/blink/web_tests/external/wpt/mathml/relations/html5-tree/a-type-getter-001-expected.txt

This is a testharness.js-based test.
[FAIL] Test anchor's type getter when empty
  assert_equals: expected (string) "" but got (undefined) undefined
Harness: the test ran to completion.
